requirements.txt
annotated-types==0.7.0
anyio==4.5.2
boto3==1.35.63
botocore==1.35.63
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.4
h11==0.14.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
